+++ examples/gcm.rb
@@ -16,7 +16,7 @@ pusher = RubyPushNotifications::GCM::GCMPusher.new "Your app's GCM key"
 # pusher = RubyPushNotifications::GCM::GCMPusher.new "Your app's GCM key", { open_timeout: 10, read_timeout: 10 }
 
 # Manually set GCM URL (e.g. useful for load testing)
-# pusher = RubyPushNotifications::GCM::GCMPusher.new "Your app's GCM key", { gcm_url: "https://example.com/gcm/send" }
+# pusher = RubyPushNotifications::GCM::GCMPusher.new "Your app's GCM key", { url: "https://example.com/gcm/send" }
 
 pusher.push [notification]
 p 'Notification sending results:'