+++ CHANGELOG.md
@@ -3,8 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unrealeased] - [unrealease date]
-- **Added**: Option (`gcm_url`) to set custom GCM endpoint (Android)
-- **Added**: Option (`apns_host`) to set custom APNS environment (iOS)
+- **Added**: Option (`url`) to set custom GCM endpoint (Android)
+- **Added**: Option (`host`) to set custom APNS environment (iOS)
 
 ## 1.0.1 - 2017-01-12
 - **Fixed**: Loading issue when `'forwardable'` wasn't previously required.